requirements.txt
# AI-Onboarder Requirements
# Python 3.10+

# Web Framework
streamlit>=1.28.0

# AI/ML
google-genai

# Git Operations
gitpython>=3.1.40

# HTTP Requests
requests>=2.31.0

# Audio/Video (Optional - for video generation feature)
gtts
pillow
moviepy
pydub
ffmpeg-python
# Utilities
python-dotenv>=1.0.0
